[NH-016] - #16 - Refactoring
